# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go backend with Ethereum tx queue and REST endpoints.",
        "title": "Go Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/tx/send": {
            "post": {
                "description": "通过 Redis 队列异步发送交易，支持 Idempotency-Key 幂等；立即返回 task_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tx"
                ],
                "summary": "发送以太坊交易（入队）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "幂等键",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "交易请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TransactionSendRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "{task_id, status}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tx/{hash}/status": {
            "get": {
                "description": "根据交易哈希查询区块链交易状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tx"
                ],
                "summary": "查询交易状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "交易哈希",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{hash, status}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tx/{taskID}": {
            "get": {
                "description": "根据 taskID 查询交易任务的当前状态与交易哈希",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tx"
                ],
                "summary": "查询交易任务状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务ID",
                        "name": "taskID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{task_id, tx_hash, status}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.TransactionSendRequest": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "gas_limit": {
                    "type": "string"
                },
                "gas_price": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}